+++ spec/spec_helper.rb
@@ -6,6 +6,7 @@ require "ast/merge/rspec"
 
 # Internal ENV config
 require_relative "config/debug"
+require_relative "config/tree_haver"
 
 # Config for development dependencies of this library
 # i.e., not configured by this library